Refactor search into separate component folder
