+++ src/components/search/CategoryFilter.tsx
