+++ src/components/search/SchoolFilter.tsx
